Create Game_Theory_Voting_System.cpp
While playing Multiplayer Game, I have analyzed the ranking framework, which will help me to achieve the desired voting system. However, many countries like New Zealand, Ireland, Australia, Malta, Slovenia, and Nauru use Ranking procedures in their democratic framework to get effective decisions and a fair winner party. So, in a desire to digitalize India's voting system, we need an upgrade. In this project, We have proposed a voting system based on Game Theory that will enable the concerned authorities to find who will be the triumphant contender for the political decision, as indicated by a particular political situation. In addition, the primary center is to Upgrade the System of Voting entirely and bring in the most crucial phase in this system of Elections for the resident and their favored chief likewise and to decrease election violence which would be done with the help of GT fragments for Rank.

Language Use C++